requirements.txt
discord.py
fade
requests
